requirements.txt
discord==2.3.2
python-dotenv
requests
Flask==2.3.3
